Remove unnecessary casts (S1905)
Drop 18 redundant casts. Each was verified safe: every '(long) i * X'
multiplication has a long right operand (X is increment, PTR_SIZE,
ENUM_SERVICE_STATUS_PROCESS_SIZE, ADDRESS.byteSize(), etc.), so the product was
already computed in long and no narrowing/overflow was being prevented. The two
'(int)' casts on short/byte values are redundant because the '&' operator
already promotes to int, and getPartitions() is already declared List<HWPartition>.
No behavior change.

+++ oshi-core-ffm/src/main/java/oshi/driver/linux/WhoFFM.java
@@ -109,7 +109,7 @@ public final class WhoFFM {
 
             try {
                 // Read array of string pointers
-                MemorySegment ptrArray = sessions.reinterpret((long) count * ADDRESS.byteSize(), arena, null);
+                MemorySegment ptrArray = sessions.reinterpret(count * ADDRESS.byteSize(), arena, null);
                 for (int i = 0; i < count; i++) {
                     MemorySegment sessionIdPtr = ptrArray.getAtIndex(ADDRESS, i);
                     if (sessionIdPtr.equals(MemorySegment.NULL)) {
@@ -127,7 +127,7 @@ public final class WhoFFM {
                 }
             } finally {
                 // Free each string, then the array
-                MemorySegment ptrArray = sessions.reinterpret((long) count * ADDRESS.byteSize(), arena, null);
+                MemorySegment ptrArray = sessions.reinterpret(count * ADDRESS.byteSize(), arena, null);
                 for (int i = 0; i < count; i++) {
                     MemorySegment ptr = ptrArray.getAtIndex(ADDRESS, i);
                     if (!ptr.equals(MemorySegment.NULL)) {
